Update corne.keymap | giving home shift another chance
if I were to switch completely to thumb shift, it would mean chiffre style boards are almost unusable for me. this is because I usually combine the two innermost thumb keys. however, with shift/esc, I would have to fit 3 functions on one key which is not possible for me considering that both symbols layer key and shift need to be used smoothly in the flow of typing